fix types

+++ snuba/datasets/entities/discover.py
@@ -260,7 +260,7 @@ TRANSACTIONS_COLUMNS = ColumnSet(
         (
             "spans",
             Nested(
-                [("op", String()), ("group", UInt(64)), ("exclusive_time", Float(64))]
+                [("op", String()), ("group", UInt(64)), ("exclusive_time", Float(32))]
             ),
         ),
     ]
